% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLM_bate_smooth_only.R, R/FLM_fixed_model.R
\name{flm_beta_smooth_only}
\alias{flm_beta_smooth_only}
\title{Title}
\usage{
flm_beta_smooth_only(
  pheno,
  mode = "Additive",
  geno,
  pos,
  order,
  basis,
  covariate,
  base = "bspline",
  interaction = FALSE
)

flm_beta_smooth_only(
  pheno,
  mode = "Additive",
  geno,
  pos,
  order,
  basis,
  covariate,
  base = "bspline",
  interaction = FALSE
)
}
\arguments{
\item{pheno}{describe}

\item{mode}{describe}

\item{geno}{describe}

\item{pos}{describe}

\item{order}{describe}

\item{basis}{describe}

\item{covariate}{describe}

\item{base}{describe}

\item{interaction}{describe}
}
\value{
a value

a value
}
\description{
Title

Title
}
