% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLM_FPCA.R
\name{flm_fpca}
\alias{flm_fpca}
\title{Title}
\usage{
flm_fpca(
  pheno,
  mode = "Additive",
  geno,
  covariates,
  pos,
  kz = 30,
  kb = 30,
  smooth.cov = FALSE,
  family = "gaussian"
)
}
\arguments{
\item{pheno}{a data frame of phenotype data. This data frame should contain three columns: pedigree id, person id, and phenotype value. These three columns should be named as ped, person, and trait, respectively.}

\item{mode}{a character value specifying the discrete realization of genetic variant function. The default of “Additive” assumes additive effect of minor alleles, “Dom” assume dominant effect, and “Rec” assumes recessive effect.}

\item{geno}{a matrix of genotype data only. The number of rows should be equal to the number of individuals in pheno. The number of columns is the total number of SNPs.}

\item{covariates}{a data frame of covariate data. This data frame should contrain at least three columns: pedigree id, person id, and at least one covariate. The first two colums should be named as ped and person, respectively.}

\item{pos}{a vector of physical positions of the SNP used in geno. The length of pos should be equal to the column number of geno.}

\item{kz}{30}

\item{kb}{30}

\item{smooth.cov}{FALSE}

\item{family}{"gaussian"}
}
\value{
a list of p values of testing the genetic effect, given by likelihood ratio test, chi-squared test, and F-distributed test, respectively.
}
\description{
Title
}
