% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLM_FPCA_no_position.R
\name{flm_fpca_no_position}
\alias{flm_fpca_no_position}
\title{Title}
\usage{
flm_fpca_no_position(
  pheno,
  mode = "Additive",
  geno,
  covariates,
  kz = 30,
  kb = 30,
  smooth.cov = FALSE,
  family = "gaussian"
)
}
\arguments{
\item{pheno}{describe}

\item{mode}{"Additive"}

\item{geno}{describe}

\item{covariates}{describe}

\item{kz}{30}

\item{kb}{30}

\item{smooth.cov}{FALSE}

\item{family}{"gaussian"}
}
\value{
a value
}
\description{
Title
}
