% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLM_survival_beta_smooth_only.R
\name{flm_surv_beta_smooth_only}
\alias{flm_surv_beta_smooth_only}
\title{Title}
\usage{
flm_surv_beta_smooth_only(
  time,
  delta,
  mode = "Additive",
  geno,
  pos,
  order,
  basis,
  covariate,
  base = "bspline"
)
}
\arguments{
\item{time}{describe}

\item{delta}{describe}

\item{mode}{describe}

\item{geno}{describe}

\item{pos}{describe}

\item{order}{describe}

\item{basis}{describe}

\item{covariate}{describe}

\item{base}{describe}
}
\value{
a value
}
\description{
Title
}
