% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLM_survival_fixed_model.R
\name{flm_surv_fixed_model}
\alias{flm_surv_fixed_model}
\title{Title}
\usage{
flm_surv_fixed_model(
  time,
  delta,
  mode = "Additive",
  geno,
  pos,
  order,
  bbasis,
  fbasis,
  gbasis,
  covariate,
  base = "bspline"
)
}
\arguments{
\item{time}{a numeric vector of follow up time for right censored data.}

\item{delta}{a numeric vector of status indicator for right censored data, normally 0=no event, 1=event.}

\item{mode}{a character value specifying the discrete realization of genetic variant function. The default of “Additive” assumes additive effect of minor alleles, “Dom” assume dominant effect, and “Rec” assumes recessive effect.}

\item{geno}{a matrix of genotype data only. The number of rows should be equal to the number of individuals in pheno. The number of columns is the total number of SNPs.}

\item{pos}{a vector of physical positions of the SNP used in geno. The length of pos should be equal to the column number of geno.}

\item{order}{an integer specifying the order of b-splines basis function, which is one higher than their degree. The default of 4 gives cubic splines. If base=”fspline”, order is ignored.}

\item{bbasis}{an integer variable specifying the number of b-spine basis functions used to expand genetic effect function.}

\item{fbasis}{an odd integer variable specifying the number of Fourier spline basis functions used to expand genetic effect function.}

\item{gbasis}{an integer variable specifying the number of basis functions used to expand genetic variant function. If base=”fspline”, basis should be an odd integer.}

\item{covariate}{a data frame of covariate data. This data frame should contrain at least three columns: pedigree id, person id, and at least one covariate. The first two colums should be named as ped and person, respectively.}

\item{base}{a character value to specify basis function system to expand the genetic effect function and genetic variant function. Options are “bspline” for b-splines and “fspline” for Fourier splines.}
}
\value{
a numeric value of p values of testing the genetic effect given by likelihood ratio test.
}
\description{
Title
}
