% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFLM_beta_smooth_only.R
\name{gflm_beta_smooth_only}
\alias{gflm_beta_smooth_only}
\title{Title}
\usage{
gflm_beta_smooth_only(
  pheno,
  mode = "Additive",
  geno,
  pos,
  order,
  basis,
  covariate,
  base = "bspline",
  interaction = FALSE
)
}
\arguments{
\item{pheno}{describe}

\item{mode}{describe}

\item{geno}{describe}

\item{pos}{describe}

\item{order}{describe}

\item{basis}{describe}

\item{covariate}{describe}

\item{base}{describe}

\item{interaction}{describe}
}
\value{
a value
}
\description{
Title
}
