% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFLM_fixed_model.R
\name{gflm_fixed_model}
\alias{gflm_fixed_model}
\title{Title}
\usage{
gflm_fixed_model(
  pheno,
  mode = "Additive",
  geno,
  pos,
  order,
  bbasis,
  fbasis,
  gbasis,
  covariate,
  base = "bspline",
  interaction = FALSE
)
}
\arguments{
\item{pheno}{describe}

\item{mode}{describe}

\item{geno}{describe}

\item{pos}{describe}

\item{order}{describe}

\item{bbasis}{describe}

\item{fbasis}{describe}

\item{gbasis}{describe}

\item{covariate}{describe}

\item{base}{describe}

\item{interaction}{describe}
}
\value{
a value
}
\description{
Title
}
