% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLM_FPCA.R, R/FLM_FPCA_no_position.R
\name{Trace.Matrix}
\alias{Trace.Matrix}
\title{Title}
\usage{
Trace.Matrix(x)

Trace.Matrix(x)
}
\arguments{
\item{x}{describe}
}
\value{
a matrix

a matrix
}
\description{
Title

Title
}
