% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLM_FPCA_no_position.R
\name{flm_fpca_no_position}
\alias{flm_fpca_no_position}
\title{Title}
\usage{
flm_fpca_no_position(
  pheno,
  mode = "Additive",
  geno,
  covariates,
  kz = 30,
  kb = 30,
  smooth.cov = FALSE,
  family = "gaussian"
)
}
\arguments{
\item{pheno}{@param pheno a data frame of phenotype data. This data frame should contain three columns: pedigree id, person id, and phenotype value. These three columns should be named as ped, person, and trait, respectively.}

\item{mode}{"Additive"}

\item{geno}{a matrix of genotype data only. The number of rows should be equal to the number of individuals in pheno. The number of columns is the total number of SNPs.}

\item{covariates}{a data frame of covariate data. This data frame should contrain at least three columns: pedigree id, person id, and at least one covariate. The first two colums should be named as ped and person, respectively.}

\item{kz}{30}

\item{kb}{30}

\item{smooth.cov}{FALSE}

\item{family}{"gaussian"}
}
\value{
a list of p values of testing the genetic effect, given by likelihood ratio test, chi-squared test, and F-distributed test, respectively.
}
\description{
Title
}
