% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLM_survival_fixed_model.R
\name{flm_surv_fixed_model}
\alias{flm_surv_fixed_model}
\title{Title}
\usage{
flm_surv_fixed_model(
  time,
  delta,
  mode = "Additive",
  geno,
  pos,
  order,
  bbasis,
  fbasis,
  gbasis,
  covariate,
  base = "bspline"
)
}
\arguments{
\item{time}{describe}

\item{delta}{describe}

\item{mode}{describe}

\item{geno}{describe}

\item{pos}{describe}

\item{order}{describe}

\item{bbasis}{describe}

\item{fbasis}{describe}

\item{gbasis}{describe}

\item{covariate}{describe}

\item{base}{describe}
}
\value{
a value
}
\description{
Title
}
